fix: all recipe route

+++ modules/recipe/repositories/RecipeRepository.ts
@@ -1,21 +1,14 @@
 import { IRecipeRepository } from "../interfaces/IRecipeRepository";
 import { CreateRecipeDto } from "../dto/createRecipeDto";
-import { Recipe } from "../entities/Recipe";
+import { RecipeWithIngredients } from "../entities/Recipe";
 import { prisma } from "$lib/db";
 import { DeleteRecipeDto } from "$modules/recipe/dto/deleteRecipeDto";
 import { AppError } from "$lib/errors/AppError";
-import {
-  ListRecipesOutput,
-  IRecipeFilters,
-  GetRecipeByIdOutput,
-} from "../dto/recipeDto";
-import {
-  RecipeCategory,
-  DietType,
-  Prisma,
-} from "../../../src/generated/prisma";
+import { DietType, Prisma, RecipeCategory, UnitType, } from "../../../src/generated/prisma";
 import { UpdateRecipeDto } from "$modules/recipe/dto/updateRecipeDto";
 import { User } from "$lib/common/User";
+import { WithPagination } from "$lib/common/types/WithPagination";
+import { IRecipeFilters } from "$modules/recipe/utils/IRecipeFilters";
 
 export const RecipeRepository = (user: User): IRecipeRepository => {
   return {
@@ -23,17 +16,65 @@ export const RecipeRepository = (user: User): IRecipeRepository => {
       const menu = await prisma.recipeMeal.findFirst({ where: { recipeId } });
       return !!menu;
     },
-    async update(recipeDto: UpdateRecipeDto): Promise<void> {
+    async update(recipeDto: UpdateRecipeDto) {
       try {
+        const recipe = await this.findById(recipeDto.id);
+        if (!recipe)
+          return null;
+
+        const ingredientsToDelete = recipe.recipeIngredients.filter(
+          (child) => !recipeDto.ingredients?.some((c) => c.id === child.id),
+        );
+
+        for (const ingredientsToDeleteElement of ingredientsToDelete) {
+          await prisma.recipeIngredient.delete({
+            where: {
+              recipeId_ingredientId: {
+                recipeId: recipeDto.id,
+                ingredientId: ingredientsToDeleteElement.id,
+              }
+            }
+          });
+        }
+
+        const ingredientsToAddd =
+          recipeDto.ingredients?.filter(
+            (newMeal) =>
+              !recipe.recipeIngredients.some(
+                (existing) => existing.id === newMeal.id,
+              ),
+          ) || [];
+
+        for (const ingredientsToAdddElement of ingredientsToAddd) {
+          await prisma.recipeIngredient.create({
+            data: {
+              recipeId: recipeDto.id,
+              ingredientId: ingredientsToAdddElement.id,
+              quantity: ingredientsToAdddElement.quantity,
+              unit: ingredientsToAdddElement.unit as UnitType,
+            },
+          });
+        }
+
         const whereAppend: Prisma.RecipeWhereUniqueInput | object =
           user.role === "user" ? { apiKey: user.apiKey } : {};
 
         await prisma.recipe.update({
           data: {
             title: recipeDto.title,
+            description: recipeDto.description,
+            isPublic: recipeDto.isPublic,
+            servings: recipeDto.servings,
+            category: recipeDto.category as RecipeCategory,
+            diet: recipeDto.diet as DietType,
+            prepTimeMin: recipeDto.prepTimeMin,
+            cookTimeMin: recipeDto.cookTimeMin,
+            instructions: recipeDto.instructions,
           },
           where: { id: recipeDto.id, ...whereAppend },
         });
+
+        return this.findById(recipeDto.id);
       } catch (error) {
         console.error("An error occurred while updating recipe", error);
         throw new AppError(
@@ -49,6 +90,13 @@ export const RecipeRepository = (user: User): IRecipeRepository => {
         const whereAppend: Prisma.RecipeWhereUniqueInput | object =
           user.role === "user" ? { apiKey: user.apiKey } : {};
 
+        const recipeExist = await prisma.recipe.findUnique({
+          where: { id: recipeDto.id, ...whereAppend },
+        });
+
+        if (!recipeExist)
+          return;
+
         await prisma.recipe.delete({
           where: { id: recipeDto.id, ...whereAppend },
         });
@@ -62,31 +110,35 @@ export const RecipeRepository = (user: User): IRecipeRepository => {
         );
       }
     },
-    async save(recipeDto: CreateRecipeDto): Promise<Recipe> {
+    async save(recipeDto: CreateRecipeDto): Promise<RecipeWithIngredients> {
       try {
         const createdRecipe = await prisma.recipe.create({
           data: {
             title: recipeDto.title,
             description: recipeDto.description,
             isPublic: recipeDto.isPublic,
+            servings: recipeDto.servings,
+            category: recipeDto.category as RecipeCategory,
+            diet: recipeDto.diet as DietType,
             prepTimeMin: recipeDto.prepTimeMin,
             cookTimeMin: recipeDto.cookTimeMin,
             instructions: recipeDto.instructions,
+            ingredients: {
+              create: recipeDto.ingredients.map((ingredient) => ({
+                quantity: ingredient.quantity,
+                unit: ingredient.unit as UnitType,
+                ingredient: {
+                  connect: { id: ingredient.id },
+                },
+              })),
+            },
             apiKey: { connect: { key: user.apiKey } },
           },
+          include: { ingredients: true }
         });
 
-        for (const ingredient of recipeDto.ingredients) {
-          await prisma.recipeIngredient.create({
-            data: {
-              quantity: ingredient.quantity,
-              ingredient: { connect: { id: ingredient.id } },
-              recipe: { connect: { id: createdRecipe.id } },
-            },
-          });
-        }
-
         return {
+          id: createdRecipe.id,
           cookTimeMin: createdRecipe.cookTimeMin,
           createdAt: createdRecipe.createdAt,
           description: createdRecipe.description,
@@ -96,10 +148,27 @@ export const RecipeRepository = (user: User): IRecipeRepository => {
           servings: createdRecipe.servings,
           title: createdRecipe.title,
           updatedAt: createdRecipe.updatedAt,
-          id: createdRecipe.id,
+          instructions: createdRecipe.instructions,
+          category: createdRecipe.category,
+          diet: createdRecipe.diet,
+          recipeIngredients: createdRecipe.ingredients.map(ingredient => ({
+            id: ingredient.ingredientId,
+            quantity: ingredient.quantity,
+            unit: ingredient.unit,
+          })),
         };
       } catch (error) {
         console.log("Error saving recipe:", error);
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+        // @ts-expect-error
+        if (error.meta.cause.includes("No 'Ingredient' record")) {
+          throw new AppError(
+            "Internal Server Error",
+            "An error occurred while saving recipe",
+            "Un des ingrédients n'existe pas dans la base de données.",
+            "error",
+          );
+        }
         throw new AppError(
           "Internal Server Error",
           "An error occurred while saving recipe",
@@ -112,7 +181,7 @@ export const RecipeRepository = (user: User): IRecipeRepository => {
       limit: number,
       offset: number,
       filters: IRecipeFilters,
-    ): Promise<ListRecipesOutput> {
+    ): Promise<WithPagination<RecipeWithIngredients[], "recipes">> {
       const where: Prisma.RecipeWhereInput = {
         ...(filters.category && {
           category: filters.category as RecipeCategory,
@@ -132,19 +201,17 @@ export const RecipeRepository = (user: User): IRecipeRepository => {
           { isPublic: true },
           ...(filters.search
             ? [
-                { title: { contains: filters.search, mode: "insensitive" } },
-                {
-                  description: {
-                    contains: filters.search,
-                    mode: "insensitive",
-                  },
+              { title: { contains: filters.search, mode: "insensitive" } },
+              {
+                description: {
+                  contains: filters.search,
+                  mode: "insensitive",
                 },
-              ]
+              },
+            ]
             : []),
         ],
       };
-
-      console.dir(where, { depth: null });
 
       const [recipes, total] = await Promise.all([
         prisma.recipe.findMany({
@@ -166,24 +233,23 @@ export const RecipeRepository = (user: User): IRecipeRepository => {
         prisma.recipe.count({ where }),
       ]);
 
-      const data = recipes.map((recipe) => ({
+      const data: RecipeWithIngredients[] = recipes.map((recipe) => ({
         id: recipe.id,
         title: recipe.title,
-        description: recipe.description || "",
+        description: recipe.description,
         imageUrl: recipe.imageUrl,
-        prepTimeMin: recipe.prepTimeMin || 0,
-        cookTimeMin: recipe.cookTimeMin || 0,
+        instructions: recipe.instructions,
+        prepTimeMin: recipe.prepTimeMin,
+        cookTimeMin: recipe.cookTimeMin,
         servings: recipe.servings,
         isPublic: recipe.isPublic,
-        category: recipe.category || "OTHER",
-        diet: recipe.diet || "OTHER",
-        ingredients: recipe.ingredients.map((recipeIngredient) => ({
+        category: recipe.category,
+        diet: recipe.diet,
+        recipeIngredients: recipe.ingredients.map((recipeIngredient) => ({
           id: recipeIngredient.ingredient.id,
-          name: recipeIngredient.ingredient.name,
           quantity: recipeIngredient.quantity,
           unit: recipeIngredient.unit,
         })),
-        mealCount: recipe._count.recipeMeals,
         createdAt: recipe.createdAt,
         updatedAt: recipe.updatedAt,
       }));
@@ -198,7 +264,7 @@ export const RecipeRepository = (user: User): IRecipeRepository => {
       };
     },
 
-    async findById(id: string): Promise<GetRecipeByIdOutput> {
+    async findById(id: string): Promise<RecipeWithIngredients | null> {
       const recipe = await prisma.recipe.findUnique({
         where: { id },
         include: {
@@ -213,46 +279,28 @@ export const RecipeRepository = (user: User): IRecipeRepository => {
         },
       });
 
-      if (!recipe) {
-        throw new AppError(
-          "Not Found",
-          "Recipe not found",
-          "Recette non trouvée.",
-          "warn",
-        );
-      }
+      if (!recipe)
+        return null;
 
-      const recipeDetailDto = {
+      return {
         id: recipe.id,
         title: recipe.title,
-        description: recipe.description || "",
+        description: recipe.description,
         imageUrl: recipe.imageUrl,
-        prepTimeMin: recipe.prepTimeMin || 0,
-        cookTimeMin: recipe.cookTimeMin || 0,
+        prepTimeMin: recipe.prepTimeMin,
+        cookTimeMin: recipe.cookTimeMin,
         servings: recipe.servings,
         isPublic: recipe.isPublic,
-        category: recipe.category || "OTHER",
-        diet: recipe.diet || "OTHER",
-        ingredients: recipe.ingredients.map((recipeIngredient) => ({
+        category: recipe.category,
+        diet: recipe.diet,
+        recipeIngredients: recipe.ingredients.map((recipeIngredient) => ({
           id: recipeIngredient.ingredient.id,
-          name: recipeIngredient.ingredient.name,
           quantity: recipeIngredient.quantity,
           unit: recipeIngredient.unit,
-          notes: recipeIngredient.notes || "",
-          category: recipeIngredient.ingredient.category || "",
-          calories: recipeIngredient.ingredient.calories || 0,
-          proteins: recipeIngredient.ingredient.proteins || 0,
-          fats: recipeIngredient.ingredient.fats || 0,
-          carbs: recipeIngredient.ingredient.carbs || 0,
         })),
-        mealCount: recipe._count.recipeMeals,
         createdAt: recipe.createdAt,
         updatedAt: recipe.updatedAt,
         instructions: recipe.instructions,
-      };
-
-      return {
-        recipe: recipeDetailDto,
       };
     },
   };
